+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- New method `Toolchain::rustc` to execute a toolchain's `rustc`.
+
 ## [0.2.0] - 2019-09-06
 
 ### Added